Duke Leve-4: ToDo, Event, Deadline
Replace Scanner with better String split() method under Conversation runs section

Create a package ip.duke for Duke suite of files

Conform code to standards
