+++ src/main/java/ip/duke/Event.java
@@ -1,0 +1,14 @@
+package ip.duke;
+
+public class Event extends Deadline {
+
+    protected Event(String description, String at) {
+        super(description, at);
+        setId("E");
+    }
+
+    @Override
+    public String toString() {
+        return String.format("%s (at: %s)", getDescription(), getBy());
+    }
+}
