+++ src/main/java/ip/duke/Todo.java
@@ -1,0 +1,9 @@
+package ip.duke;
+
+public class Todo extends Task {
+
+    protected Todo(String description) {
+        super(description);
+        setId("T");
+    }
+}
